Completed the basic functionality of this program.
Modified the function prototype of the function that is called by the
updating sweep such that it accepts: (a) the index of the sub-image to use
for updating, (b) the x,y location of the corner where the sub-image is to
be placed, and (c) the width,height of the sub-image. Modified the original
"nothing_func()" to have the appropriate prototype. (This function now
accepts the integer values mentioned in additionto the payload void*
argument.) The function was modified to properly position the sub-image in
the window that mirrors the X server that is being leached.

The function prototype of the updating-sweep function was modified to be
compatible with the new functionality. This function was also modified in
how it makes the call to the function provided via the function-pointer.
IN 2020/09/30 "master"